refactor: use the most accurate pod spec to compute differences
The PodSpec template defined by a controller might not represent the final
spec of the pods. For example, a LimitRanger controller could change the spec
of new pods after their creation to set default resource requests and limits.
To ensure that we have a reliable comparison source, we have no choice but to
list the pods and find those who are dependents of the controller to get the
most up-to-date spec.