fix: correct the computation of the percentage differences and stats
Fix the computation of the CPU/Memory difference columns to express
the result as the increase/decrease of the request in terms of the
recommendation.

Print the memory recommendation quantity using the same scale as the
request in a human-friendly format.

Scale the statistics values to the number of replicas requested in
the spec of each controller.